added test categories cards
