+++ components/View profile.js
@@ -12,9 +12,10 @@ import { SvgUri , SvgCssUri } from 'react-native-svg';
 import * as yup from 'yup';
 import { Alert, I18nManager, TouchableOpacity, Image } from 'react-native';
 import getAuthData from '../hooks/getAuthData';
+import Lightbox from 'react-native-lightbox-v2';
 
   //address is still to be added
-export default function Viewprofile({navigation,role,Address_label,Name_label,header_color}){
+export default function Viewprofile({navigation,role,Address_label,Name_label,header_color,navigator}){
     const {t,i18n} = useTranslation();
     const [loading , Isloading] = useState(true);
     const [profileData,setProfileData] = useState();
@@ -154,15 +155,11 @@ export default function Viewprofile({navigation,role,Address_label,Name_label,he
                 <VStack  justifyContent='center'>
                     <VStack h='25%'>
                         <HStack w='100%' h='50%'  bgColor={header_color} justifyContent='center' >
-                                {/* <HStack  style={{aspectRatio:1}} w='auto' borderRadius={50}  mt={12} zIndex={3} justifyContent='center' >
-                                    <SvgUri width='200' height='120' uri={profileData.picture}/>
-                                </HStack> */}
-                                <Avatar size='xl' source={{uri:profileData.picture}} zIndex={3} mt='5' />
+                            <Lightbox navigator={navigator} ><Avatar size='xl' source={{uri:profileData.picture}} zIndex={3} mt='5' /></Lightbox>
                         </HStack>
                         <HStack  h='54%' justifyContent='center'>
                             <Text  fontWeight='bold' mt='12'  fontSize='2xl'> {profileData.name}</Text>
                         </HStack>
-                        {/* <Text textAlign='center' color='danger.900'>{t('You can edit any field of your Personal Data by just clicking on it and filing your new data 😊')}</Text> */}
                     </VStack>
                     <ScrollView mt={5}  h='60%'>
