+++ package.json
@@ -40,6 +40,7 @@
     "react-i18next": "^11.12.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-42.0.0.tar.gz",
     "react-native-gesture-handler": "~1.10.2",
+    "react-native-lightbox-v2": "^0.9.0",
     "react-native-ratings": "^8.1.0",
     "react-native-reanimated": "~2.2.0",
     "react-native-restart": "^0.0.24",
