+++ screens/Doctors/View Doctor Profile.js
@@ -7,7 +7,7 @@ import Viewprofile from '../../components/View profile';
 const ViewDoctorProfile=({navigation})=>{
     return(
         <NativeBaseProvider>
-            <Viewprofile navigation={navigation} role='doctor' Address_label='Clinic Address' Name_label='Doctor Name' header_color='#003049' />
+            <Viewprofile navigation={navigation} navigator={navigator} role='doctor' Address_label='Clinic Address' Name_label='Doctor Name' header_color='#003049' />
         </NativeBaseProvider>
     )
 }
