+++ screens/Users/View User Profile.js
@@ -5,7 +5,7 @@ import Viewprofile from '../../components/View profile';
 const ViewUserProfile=({navigation})=>{
     return(
         <NativeBaseProvider>
-            <Viewprofile navigation={navigation} role='user' Address_label='Address' Name_label='Username' header_color='success.500' />
+            <Viewprofile navigation={navigation} navigator={navigator} role='user' Address_label='Address' Name_label='Username' header_color='success.500' />
         </NativeBaseProvider>
     )
 }
